# frontend/openapi-legacy.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Incident Investigation Console API",
    "version": "0.1.0"
  },
  "paths": {
    "/ui/cases/{case_id}/summary": {
      "get": {
        "summary": "Get Summary",
        "operationId": "get_summary_ui_cases__case_id__summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UIBasicCase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ui/cases/{case_id}/frame": {
      "get": {
        "summary": "Get Frame",
        "operationId": "get_frame_ui_cases__case_id__frame_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationFrame"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ui/cases/{case_id}/evidence-page": {
      "get": {
        "summary": "Get Evidence Page",
        "operationId": "get_evidence_page_ui_cases__case_id__evidence_page_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "frame_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Frame Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 40,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ui/cases/{case_id}/evidence-items/{evidence_id}": {
      "get": {
        "summary": "Get Frame Evidence",
        "operationId": "get_frame_evidence_ui_cases__case_id__evidence_items__evidence_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evidence Id"
            }
          },
          {
            "name": "frame_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Frame Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UIEvidence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ui/cases/{case_id}": {
      "get": {
        "summary": "Get Case",
        "operationId": "get_case_ui_cases__case_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UICase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ui/cases/{case_id}/evidence": {
      "get": {
        "summary": "Get Evidence",
        "operationId": "get_evidence_ui_cases__case_id__evidence_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UIEvidenceList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ui/cases/{case_id}/hypotheses": {
      "get": {
        "summary": "Get Hypotheses",
        "operationId": "get_hypotheses_ui_cases__case_id__hypotheses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UIHypothesisGraph"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ui/cases/{case_id}/reasoning-context": {
      "get": {
        "summary": "Get Context",
        "operationId": "get_context_ui_cases__case_id__reasoning_context_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasoningContextSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CaseStatus": {
        "type": "string",
        "enum": [
          "NEW",
          "INVESTIGATING",
          "WAITING",
          "ESCALATED",
          "CLOSED",
          "CLOSED_VERIFIED"
        ],
        "title": "CaseStatus"
      },
...
      "EvidencePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UIEvidence"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "eligibility_denied_count": {
            "type": "integer",
            "title": "Eligibility Denied Count"
          },
          "frame_id": {
            "type": "string",
            "title": "Frame Id"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "eligibility_denied_count",
          "frame_id"
        ],
        "title": "EvidencePage"
      },
...
      "Freshness": {
        "type": "string",
        "enum": [
          "CURRENT",
          "STALE",
          "UNKNOWN"
        ],
        "title": "Freshness"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestigationFrame": {
        "properties": {
          "frame_id": {
            "type": "string",
            "title": "Frame Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "case_revision": {
            "type": "string",
            "format": "date-time",
            "title": "Case Revision"
          },
          "evidence_fingerprint": {
            "type": "string",
            "title": "Evidence Fingerprint"
          },
          "registry_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registry Version"
          },
          "route_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Revision"
          },
          "latest_agent_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Agent Run"
          },
          "assembled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assembled At"
          },
          "route_summary": {
            "items": {
              "$ref": "#/components/schemas/TraceField"
            },
            "type": "array",
            "title": "Route Summary",
            "default": []
          },
          "investigation_allowed": {
            "type": "boolean",
            "title": "Investigation Allowed"
          },
          "current_evidence_total": {
            "type": "integer",
            "title": "Current Evidence Total"
          },
          "case_summary": {
            "$ref": "#/components/schemas/UICase"
          },
          "financial_truth": {
            "items": {
              "$ref": "#/components/schemas/FinancialTruth"
            },
            "type": "array",
            "title": "Financial Truth"
          },
          "financial_identity": {
            "$ref": "#/components/schemas/PaymentIdentityContext"
          },
          "financial_identity_dimensions": {
            "items": {
              "$ref": "#/components/schemas/TraceField"
            },
            "type": "array",
            "title": "Financial Identity Dimensions"
          },
          "current_evidence": {
            "items": {
              "$ref": "#/components/schemas/FactCapsule"
            },
            "type": "array",
            "title": "Current Evidence"
          },
          "hypotheses": {
            "$ref": "#/components/schemas/UIHypothesisGraph"
          },
          "gap_capabilities": {
            "items": {
              "$ref": "#/components/schemas/GapCapability"
            },
            "type": "array",
            "title": "Gap Capabilities"
          },
          "evidence": {
            "$ref": "#/components/schemas/EvidencePage"
          },
          "timeline": {
            "$ref": "#/components/schemas/TracePage"
          },
          "planner_trace": {
            "$ref": "#/components/schemas/TracePage"
          },
          "tool_trace": {
            "$ref": "#/components/schemas/TracePage"
          },
          "registry_source_trace": {
            "$ref": "#/components/schemas/TracePage"
          },
          "route_trace": {
            "$ref": "#/components/schemas/TracePage"
          },
          "knowledge_retrieval_trace": {
            "$ref": "#/components/schemas/TracePage"
          },
          "work_trace": {
            "$ref": "#/components/schemas/TracePage"
          },
          "side_effect_trace": {
            "$ref": "#/components/schemas/TracePage"
          },
          "recovery_trace": {
            "$ref": "#/components/schemas/TracePage"
          },
          "evaluation_trace": {
            "$ref": "#/components/schemas/TracePage"
          },
          "closure_trace": {
            "$ref": "#/components/schemas/TracePage"
          },
          "notice": {
            "type": "string",
            "title": "Notice",
            "default": "READ ONLY \u00b7 APPLIED \u2260 VERIFIED \u00b7 Historical guidance is not current Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "frame_id",
          "case_id",
          "case_revision",
          "evidence_fingerprint",
          "registry_version",
          "route_revision",
          "latest_agent_run",
          "assembled_at",
          "investigation_allowed",
          "current_evidence_total",
          "case_summary",
          "financial_truth",
          "financial_identity",
          "financial_identity_dimensions",
          "current_evidence",
          "hypotheses",
          "gap_capabilities",
          "evidence",
          "timeline",
          "planner_trace",
          "tool_trace",
          "registry_source_trace",
          "route_trace",
          "knowledge_retrieval_trace",
          "work_trace",
          "side_effect_trace",
          "recovery_trace",
          "evaluation_trace",
          "closure_trace"
        ],
        "title": "InvestigationFrame"
      },
...
      "UIBasicCase": {
        "properties": {
          "case_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_:@/.-]*$",
            "title": "Case Id"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "case_id",
          "status",
          "updated_at"
        ],
        "title": "UIBasicCase"
      },
      "UICase": {
        "properties": {
          "case_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_:@/.-]*$",
            "title": "Case Id"
          },
          "internal_order_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_:@/.-]*$",
            "title": "Internal Order Id"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "budget": {
            "$ref": "#/components/schemas/CaseBudget"
          },
          "financial_subject": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialSubject"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "case_id",
          "internal_order_id",
          "status",
          "updated_at",
          "budget",
          "financial_subject"
        ],
        "title": "UICase"
      },
      "UIEvidence": {
        "properties": {
          "claim_type": {
            "$ref": "#/components/schemas/ClaimType"
          },
          "value": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          },
          "subject": {
            "$ref": "#/components/schemas/FactSubject"
          },
          "business_time": {
            "type": "string",
            "format": "date-time",
            "title": "Business Time"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "completeness": {
            "$ref": "#/components/schemas/Completeness"
          },
          "source": {
            "$ref": "#/components/schemas/FactSource"
          },
          "protocol_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9]+([._+-][A-Za-z0-9]+)*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol Version"
          },
          "evidence_refs": {
            "items": {
              "type": "string",
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_:@/.-]*$"
            },
            "type": "array",
            "title": "Evidence Refs"
          },
          "evidence_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_:@/.-]*$",
            "title": "Evidence Id"
          },
          "observation_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_:@/.-]*$",
            "title": "Observation Id"
          },
          "event_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Time"
          },
          "strength": {
            "$ref": "#/components/schemas/EvidenceStrength"
          },
          "extractor_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9]+([._+-][A-Za-z0-9]+)*$",
            "title": "Extractor Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "claim_type",
          "value",
          "subject",
          "business_time",
          "observed_at",
          "freshness",
          "completeness",
          "source",
          "protocol_version",
          "evidence_refs",
          "evidence_id",
          "observation_id",
          "event_time",
          "strength",
          "extractor_version"
        ],
        "title": "UIEvidence"
      },
      "UIEvidenceList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UIEvidence"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "eligibility_denied_count": {
            "type": "integer",
            "title": "Eligibility Denied Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "eligibility_denied_count"
        ],
        "title": "UIEvidenceList"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}